requirements.txt
langchain
streamlit
openai-whisper
sounddevice
numpy
torch
deepseek-coder
transformers
sentence-transformers